+++ include/mbgl/style/expression/expression.hpp
@@ -173,7 +173,9 @@ enum class Kind : int32_t {
     ImageExpression,
     In,
     Within,
-    Distance
+    Distance,
+    IndexOf,
+    Slice
 };
 
 class Expression {